Move ToolStripPanelCellToControlEnumerator to a separate file
